thumbnailer: Use the correct typecast
A gint64 is being cast to a gint.

Fallout from e62947cd9c7a8cfc8a040e2111be4b5fc2d71cbf

https://bugzilla.gnome.org/show_bug.cgi?id=690255

+++ src/photos-thumbnailer.c
@@ -425,8 +425,8 @@ photos_thumbnailer_generate_thumbnail_pipeline (GObject *source_object, GAsyncRe
         }
       else
         {
-          load_height = (gint64) data->original_height;
-          load_width = (gint64) data->original_width;
+          load_height = (gint) data->original_height;
+          load_width = (gint) data->original_width;
         }
     }
   else
